refactor: move `VmRef` to separate file
